# Conflicts:
#	src/main/kotlin/com/example/Application.kt
#	src/main/kotlin/com/example/db/DatabaseSingleton.kt
#	src/main/kotlin/com/example/modules/address/controller/AddressRoutes.kt
#	src/main/kotlin/com/example/modules/article/controller/ArticleRoutes.kt
#	src/main/kotlin/com/example/modules/article/model/Article.kt
#	src/main/kotlin/com/example/modules/client/ClientService.kt
#	src/main/kotlin/com/example/modules/order/controller/OrderRoutes.kt
#	src/main/kotlin/com/example/modules/order/dao/OrderDAOFacadeImpl.kt
#	src/main/kotlin/com/example/modules/product/controller/ProductRoutes.kt
#	src/main/kotlin/com/example/modules/product/dao/ProductDAOFacade.kt
#	src/main/kotlin/com/example/modules/product/dao/ProductDAOImpl.kt
#	src/main/kotlin/com/example/modules/product/model/CreateProductDTO.kt

+++ src/main/kotlin/com/example/modules/product/controller/ProductRoutes.kt
@@ -1,7 +1,7 @@
-@file:Suppress("ktlint:standard:no-wildcard-imports")
-
 package com.example.modules.product.controller
 
+import com.example.modules.address.dao.addressDao
+import com.example.modules.address.model.Address
 import com.example.modules.product.dao.productDao
 import com.example.modules.product.model.CreateProductDTO
 import com.example.modules.product.model.Product
@@ -12,27 +12,44 @@ import io.ktor.server.routing.*
 import io.ktor.server.util.*
 
 fun Route.product() {
-    get("/") {
-        call.respondRedirect("product")
-    }
-
-    route("product") {
+    route("/product") {
+        get {
+            call.respond(productDao.allProducts())
+        }
+        get("/{id}") {
+            val id = call.parameters["id"]?.toIntOrNull()
+            if (id == null) {
+                call.respondText("Invalid ID")
+            } else {
+                val product = productDao.product(id)
+                if (product == null) {
+                    call.respondText("Product not found")
+                } else {
+                    call.respond(product)
+                }
+            }
+        }
         post {
-            val product = call.receive<CreateProductDTO>()
-            val createdProduct = productDao.addNewProduct(product)
-            call.respondRedirect("/product/${createdProduct?.id}")
+            val productDTO = call.receive<CreateProductDTO>()
+            val createdProduct = productDao.addNewProduct(productDTO)
+            if (createdProduct == null) {
+                call.respondText("Failed to create product")
+            } else {
+                call.respond(createdProduct)
+            }
         }
-
-        get("{id}") {
-            val id = call.parameters.getOrFail<Int>("id").toInt()
-            val product: Product = productDao.product(id)!!
-            call.respond(product)
-        }
-
-        delete("{id}") {
-            val id = call.parameters.getOrFail<Int>("id").toInt()
-            val deleted = productDao.deleteProduct(id)
-            call.respond(deleted)
+        delete("/{id}") {
+            val id = call.parameters["id"]?.toIntOrNull()
+            if (id == null) {
+                call.respondText("Invalid ID")
+            } else {
+                val deleted = productDao.deleteProduct(id)
+                if (deleted) {
+                    call.respondText("Product $id deleted")
+                } else {
+                    call.respondText("Product not found")
+                }
+            }
         }
     }
 }
